Commands & handlers fot operation on payments
You should notice that `with_aggregate` has been abandoned here, and
replaced with new `rehydrate` method. Rehydrate is a term used in event
sourcing and meaning apply all past domain events to new aggregate.
I do not want to use `with_aggregate` here because it forces the
aggregate to have a single argument constructor with aggregate id.
Rehydrate does not make any assumption on how the aggregate object is
created. It simply needs a empty (with initial state) aggregate object
and a name of a stream where it's domain events are stored.