Back to green mutant.
Previous attempt to Order/Cart refactoring was based on the strategy of duplicated flow for the time being. But it had other drawbacks, mutation coverage being just a minor problem.

The incorrect assumption was to reuse card is as order id. This could make the transition easier but it led to artificial testing.