Moved cmd handlers to one file ordering.rb
It's mostly subjective, but now that we're going to change many cmd handlers at once (when we move logic to Cart) it feels this belongs together.

Zeitwerk allows it, which is cool.

Potential next steps here could be making this one class instead of multiple classes. However, this would require wrapping the method calls in command handlers configuration with lambdas.