Strict Ruby version specification to make everyone happy and sad at the same time.
Pro:
* most popular Ruby version manager does not display an error as the
  very first thing you see in the project (<3 RVM)

Con:
* bundler would prevent anyone from using otherwise perfectly fine Ruby
  2.7.0, 2.7.1, 2.7.3 and 2.7.9999 in development,
  one may argue this is a Good Thing, I'd say that adds unecessary
  friction (I assume CI and deployment Ruby versions are specified
  strictly)

Unfortunately there's no other way to tell Heroku what Ruby version to
use than in Gemfile. So a deployment target dictates that (we're
downstream).

Since we have to have this version specified in Gemfile and at the same time
don't want to make it harder for newcomers (RVM cannot into non-strict
version specification), let's assume in this context "worse is better"
and make it strict in development too.