Removed missing customer order check
At this moment it's a dead code.
We ensure at the cmd level if customer id can be coerced to an integer.

This could be tested at the level of aggregate only, but it doesn't sound good.

It would be nice to have a feature of ensuring whether a customer exists, but that's a separate story.

Thanks to that, we can improve mutant coverage (remove one ignore).