Removed a test which mutation-wise was redundant.
This code was already covered by other tests, almost at the same level of abstraction. The other tests just test the consequences, while this one asserted events.

It's an interesting dilemma how to approach redundant tests - I'm still not sure so this is more an experiment.

This test was covering a command handler which can't be really unit tested at the aggregate/unit level.