Payment Gateway  knows about order value to authorize
For now, we use a FakeGateway (configurable in environment).
AuthorizePayment cmd handler has now coupling to Orders read model - which is a bad thing and needs to be fixed.

The way to go is to now refactor the codebase to extract a Pricing module and let Pricing calculate the value and then somehowe communicate it the value to Payments.