A simple Pricing BC introduced
SetPrice command + PriceSet event
ProductsCatalog read model wired to update the price
Told Zeitwerk to ignore Pricing - this is simple enough to keep in 1 file for now.
Mutant works but ignores call - due to stream_name not really needing the product_id as part of its name.
The Pricing::Product aggregate is probably an overkill for now. I was thinking about a ProductPrice aggregate with "activate" method. Maybe an option for later.