Removed the concept of Payment transaction_id
It was a nice concept but not really used in any feature yet. Removing this simplifies the connection between Pricing and Payment.

Maybe in the future we might have a proper Transaction domain object in Payments.

Right now Payments is 1-1 with an Order. In later featues we might want to split it to allow for one payment to pay for many orders or for many payments per one order.