Pricing now calculates total order value
This value is then published as an event and then consumed in the Payments area.
Thanks to that Payments no longer couples to Orders read model - technical debt cleaned up.
At the moment the total value is calculated after order is submitted. We'll calculate it on every order change soon.