Extracted CRM configuration
Which now results in all BC and read models extracted from main configuration file. This makes sense, as their details are like private.

The only things which are left in the global glue, are:

- BC declaration
- Processes
  - ReleasePaymentProcess
  - OrderConfirmation
- event handlers (we can argue they are processes too)
  - when order submitted, calculate total value
  - when value calculated, tell it to Payments

Processes are cross-BC. We can assume that event handlers are process managers (special case with just one event).

Whatever event is used by a process, it makes the event "public". Other events can treated as private to their BC.