More explicit lib/ requires
When I change cache_classes to true in dev mode (tempting) then those requires are not needed. This is unexpected for me, as zeitwerk is set to ignore this directory. For now I prefer explicit requires. But the main vision is completely get rid of zeitwerk and any other code loaders - in this project we need a very granular control over what is loaded.