Design change - discount details are not required here
This is an interesting suggestion from mutant. While we do want to know the state changes between order and discounted order, we don't really need to discount details at this leve.
I guess, they will be needed once we add publishing events. Which may or may not be at this level.
My plan is to publish events from the command handler, avoiding the aggregate root pattern here.