Payments now use discounted value
The way it works is that internally Pricing when PercentageDiscountSet happens it triggers a command to Pricing to calculate total value. I'm not confident in this solution, but seems to work fine.
From the outside it's a black box.

I've had an issue with mutant here - how to test that we need the last event from a stream?

BTW, it's yet another situation where we do event sourcing right-fold, just taking the last relevant event to build the right object - a mix of event sourcing and factory. I like the name of right-fold for that - kudos to @pawelpacana for this name.