Fixed discount bug
The bug showed up because there's a difference between the Inmemory implementation and the real one (JSON serializer treating decimals as strings).
The solution was to be explicit with the event schema, which takes care of transforming the data.

This commit also introduces the maintaining of inmemory/res adapters in tests to easily switch between them.

Some tests had to be now extended with cleaning up some db tables because they were now using the real db.