Simplified Orders read model - receive calculated values
It's usually a challenge to decide on how much the read model should calculate and what should be left to the write side.

In this case, now that we have discounts logic, it makes sense to keep the read models unaware of calculation logic.

What I'm not sure about is whether we should go for separated events per each value or such summary events like OrderTotalValueCalculated.