Simplified Orders read model
Moved event handlers to one place, which allowed for some code duplication detection and removal.
This refactoring is made as part of the feature where we will group this read model events as one stream. This allows for easier read model rebuilds and also for benefitting from RES browser feature (show order history as events).