None were used so far
The Rails default of "include everything" works great at the beginning.
In the long term, we'd rather not bother with dead dependencies, dead
code and keeping them up-to-date-and-ready for no use.

Additionally the capybara dependency broke mutant (regexp gem) while
literally no added value.

Happy to re-add when really needed.