Rails --minimal
Only the very basic frameworks we use and the gems to support it.

In the beginning of this app we thought that having a stock,
everything-included Rails app was the best choice for a sample
application. Even having Spring enabled by default, having very much
against it. Such app differed from a regular Rails one with only
the pieces we have introduced.

Yet it brought much of the noise in code discovery via browsing
and some burden in maintaining not-used functionalities of the framework
over the years. Frameworks bits come and go (sprockets, turbolinks,
weof the framework). Configuration samples and conventions change. And
we pay attention to it while making no use of it.

Staying small and nimble lets us focus on important parts, not being
distracted. When needed, we can always enable the parts of the
framework.