Introduce CustomerRepository to eventually detach from persistence details
In order to drop ActiveRecord dependency, we first introduce repository
pattern (or simply a database adapter) here.

An application, bringing its persistence library of choice (AR in Rails)
would then need to provide specific implementation.

In CRM context it would be enough to rely on in-memory implementation
and skip dealing with databases, migrations and whatnot.