Dumb code over meta
Metaprograming can be fun. It's tempting to "reduce boilerplate" with it.
Unfortunately at the same time it makes reading code worse and debugging
it a little bit harder.

Let's admire the power of metaprogramming, programmer's ability to use
it and at the same time downgrade this part to dumbcode for future readers
and Rubymine code explorers.