PriceSet no longer goes through Catalog
In the current features we don't have a need (nor any logic) to
justify the flow where Catalog knows about the price.

I can imagine that in the future it might be possible.

As of now, Catalog is going to be just for the Product registration
logic.

Thanks to this change, Catalog no longer needs to know about ProductRepository.

A reminder that we have Products as a read model - obviously this one is interested in the price changes.

This commit is part of a bigger refactoring - to have Catalog and CRM writes decoupled from reads.