Subscriber orders read model to product events
This is a step towards decoupling this read model from Products read model.
An interesting thing here - we link events within a stream and we use order_id as part of the stream name. With those product events this no longer works.

Having a stream per a read model is very useful for rebuilding.
open question: should we link those product events too?