Orders read model uses its own Products lookup table
Removed dependency to ProductRepository.
This approach is highlighting the autonomous aspect of read models.
There are some downsides - keeping a copy of all products might be a big cost for some situations.

Mutant ignore line added - deferred this to solve later. It's better to change the way this read model is unit tested. Now it's more of an integration test.

One idea during working on this all - keep view and schema together with the read model. Worth trying at some point.