Extracted Processes::ProcessManager from OrderConfirmation
This is an experiment to have a better abstraction to represent
process managers in our codebase (and possibly to move it to RES).

It's more like an exploration so far, feel free to change it.

The main idea is to simplify the creation of process managers.
Get rid of the need for the typical boilerplate.

Thanks to this simplification, it was easier to deal with mutant
and increase the coverage here.