Revert "#57 - added a possibility to archive an order"
Thanks to @lukaszreszke for starting this and asking right questions.
This feature was meant to solve a UI problem. That problem could
be better solved via some kind of filtering.
Implementing this at the domain level seems not right, after some
thinking.
Also, seeing the new state in Order triggered me into finding some
simplifications - the Order states are a bit artificial and could
be removed.
Order became like a coordinator or centralizer of other BCs - this
doesn't feel right. We'll try a more decentralized approach and see
how that will feel.