Replaced RegisterCommand with http call in integration tests
In ideal world integration tests shouldn't use commands. Commands are hidden as the black box in the application.
Integration tests are the top of the testing pyramid and should assume as minimal internals as possible.

We should test via the HTTP layer instead.

Sometimes, when for some reason, certain data changes are not possible via the http, then using commands is allowed.

Also, when a feature is in progress, it might be a good idea to start with commands in the integration test but over time replace with "UI simulation".