Moved order discount to be part of Pricing::Order
This makes it less cumbersome to calculate total value.
However this is also creating a "big aggregate" problem.

This is probably not the final solution but for now it's a step
to make calculating total value easier to follow.

For now, time promotions are left to be implemented "the old way".
We will turn them into a process later, probably.

This change is incompatible with old data.
The Discounts::Order events are now part of a different stream.

Remember to reset the data locally.