Moved time promotion tests to their own file
The test files are as important as production code.
We need to find ways of structuring them.

BTW, In the last refactoring I found it useful that we don't test
aggregates directly (at least not in Pricing). Thanks to that
it was less tests to change when we moved from where the events were
published.