Domain gems no longer require full RES, replaced with RubyEventStore
This results in dropping the dependencies to Active* gems.
No need for the upgrades anymore.
Within the domain gems we use InMemory for tests/development, the ActiveRecord part was not really used.

Where full RailsEventStore is now required is rails_application.
This makes sense as this is the integration/application level.