Cleaning up the client panel
We started with a sales panel first. Then we created a simple client view.
The goal is to have them separated (at the application level) in terms of controllers,
views, read models etc.
Removed links which led from clint panel to sales panel. We don't want to allow that.
Those will be used by different roles and with authorization rules later on.