Experimenting with the idea that Rails views could be closer to read models
This is a rather not successful attempt, but may work as a draft for now.
The high level goal is to improve quality/testability of our views/read models.

Right now the read model was only responsible for crafting data, without html.
It's the Rails controller which knows where to find the view file and renders it.

Ideally, I'd like to call the "rendering" machinery directly from the read model, but I failed to do it. Instead I'm passing the whole controller instance as a param just to call "render".

I also failed at moving the view file to the directory where read model lives. Rails allows appending the views path, but then I've had a problem with the layout file not being found. All help is welcome :)

I renamed :index to :list mostly so that mutant is happy. In situation where Rails is explicit it's hard to become implicit without mutant complaining. Here, we could just not pass the view file, because of conventions.

Next steps are to take the view to be covered with tests at the read model tests level. Now, we only test them at the Rails integration level which is an overkill.

It's very tempting to move to some Ruby-code-views (like Cells), where we could just have it covered with mutant easily. Other ideas?

I'm making those experiments only at the client panel level - it's simple enough here.