Refactored Products read model
in the search for more generic read model API...
It feels like 80% of read models is just copy one attribute from event to a column
edge cases:

- create record
- delete record
- nested attributes in events (like here with tax_rate[:code])
- different attribute vs column name

Possibly we can simplify the ActiveRecord declaration, table_name.
Also, once we want to support rebuilding read models, we may want to have migrations
as part of this code.