Client can add/remove items to cart and submit - first draft
At the moment the process of adding to cart looks similar to how sales people do it.
We should make it more like a cart - adding from a product page, for example.

Orders read model is "reused" from the sales panel which is a debt taken here.
Also, new/edit pages are using erb (instead of ruby views).

Those are shortcuts for now to have the cart ready.
Tailwind UI could be considered for this part later.