Deal with tests after changing to async read_model
Some of the assertions that relied on the redirection had to be removed.
The reason is the fact that now instead of redirecting we use broadcast
to update the values on the UI. There's no more redirect. So it cannot
be relied on.

Luckily mutant coverage didn't drop significantly. Http GET requests can still be
used to assert UI changes. However it's not possible (or I didn't find a
way) to assert turbo broadcasted updates.