Moved math to /ecommerce
It used to live as a top-level gem/dependency.
It's only used in 1 process manager which coordinates Invoicing.

At first I wanted to move it to the processes. But actually, a separate ecommerce/math dependency makes some sense.

The main goal is to simplify the dependencies here to make easier reusability of domains/BCs.