Simplify time promotions
This is just the first step. The domain model will be changed further.

It currently seems that time promotions are only created. The creation
process was divided into multiple commands, which complicates event
handling and building async read model, bringing no benefit at the same
time.

The direction is to convert time promotions to be just business use case
of discount, instead of having dedicated aggregate to deal with something
that the other (discount) model is capable of doing.

The reason is mostly the fact that if we ever want to introduce other
kind of discount, we probably wouldn't like to introduce new domain
model. We want to reuse those pieces.