Script to fix broken time promotion
TimePromotionCreated event was extended by attributes from
TimePromotionLabeled, TimePromotionDiscountSet and TimePromotionRangeSet
events.

This approach is not the best, as it makes it way harder to perform a
migration.

There's only one broken TimePromotionCreated event (luckily). It lacks
the data that is expected, hence it cannot be loaded (dry-schema kicks
in) and it throws 500s.

This script is based on the fact that there is only 1 event and this is
not recommendation to follow blindly. I am aware that I am modyfing
event's history. Such approach is not desired in real production
environment unless absolutely necessary.

This is a good lesson to plan event migration a little bit better next
time :)

Sources:

https://blog.arkency.com/4-strategies-when-you-need-to-change-a-published-event/
https://railseventstore.org/docs/v2/migrating_messages/