Packwerk setup
After initial research and experiments it seems to be promising for our needs.

In this project we're trying to really secure the boundaries:

- between read models
- between BCs
- between processes

and others.

Packwerk seems a bit invasive with the package.yml file everywhere. Maybe there are other ways. I'd prefer one configuration file to define the rules.

In the future we can consider adding it to CI.
For now I want it to part of the project to keep experimenting.

I wasn't able to make Packwerk show me the violation in the Customers read model reaching to ClientOrders. Once I can do it, we will enable the CI.

I see this tool as a hope for architectural linter.

Which should still allow shortcuts/ignores as needed and decided. Similarly like mutant.