Make reservation process synchronous
Sometimes when an order is submit and the user is redirected to the
/show view, the order state is Draft.

This happens, because the order is not really submitted. First it is
presubmitted, and then the reservation process checks whether there is
enough stock. However, with how to UI is currently designed, the state
of the order is not correct (it displays draft, as the process didn't
manage to update it).