Hide remove item button when item not possible to remove
This change applies to Sales Panel.
A similar change is needed in the Client Panel.

Tested this via integration test and checking if remove url
is visible or not in different states.
This doesn't test the Turbo mechanism which is used here too.

I don't like this duplication between the Rails view button and the same in Turbo call. I suppose we should be less granular and just rerender whole table here instead of single values.

Another ugly place is now this if/else in the view. Hard to avoid it, but maybe once we move this view to Ruby code, it can be better structured (plan for the future).