Remove sidekiq
Recently we removed async jobs from this project. This is the next step to remove sidekiq dependency.
All was good with async sidekiq implementation here, we're in the process of creating a new repo which shows async solutions. Here in this repo it will be simpler to rely on sync approaches.