Introduce a new write model alongside the old one
The idea for plugging in a new write model is that we want to begin its
stream with migration event, that's why we check whether migration event
is already in the stream

We also use an advisory lock to avoid race conditions when adding the
migration event.

New tests were added to check whether the logic of
Inventory::ProductService works correctly for all 3 cases.