Get rid of calculating totals and subtotals in Offer
Calculate and publish offer item prices in Add/Remove items events.
In addition, publish the OfferItemsPricesRecalculated event when
prices are affected by adding/removing discounts.

I also introduced the idea of PricingPolicy, currently implemented
as DefaultPricingPolicy. Its responsibility is to determine
prices of items in the context of the offer.
I hope that in the future it will allow us to inject custom policies
in configuration of the Pricing BC and get applications without
hardcoded pricing rules.