extract Infra::ProcessManager
The idea is to extract the code that is repeated in all processes.
* initialization
* subscriptions
* loading/building state
* acting

The acting part is generate/call commands based on the current state

The current state always reflects all events passed to the process
manager so far (including the event passed to current call).

ProcessManager class, which includes module should:
* define State class
* pass it to include directive, e.g.:
  `include Infra::ProcessManager.with_state(ProcessState)`
* define `process_id(event)` method. It should extract process
  identifier from event payload
* define `act` method. It's responsibility it to generate
  and call commands based on the current state