Publish OrderTotalValueCalculated from CalculateTotalValue
Instead of publishing this event from the aggregate now we're going to
publish it from the event handler. This is temporary. Ideally we'd have
an calculator for that that would subscribe to all the events that
affect the offer.

Additionally we'll get benefits from both of suggested solutions up
until now:
- we'll remove calculate_order_total_value from aggregate in next
  commits
- read models are going to subscribe to whatever is published from
  CalculateTotalValue and won't have to calculate anything on their own
- in next steps the events such as PriceItemAdded, PriceItemRemoved will
  get rid of base_total_value and total_value
- change is going to be simpler as the event stands still, it is just
  going to be produced by different component. One that will be
  responsible for doing calculations. This way we'll make Offer care
  only about items and promotions for the offer, not the price itself.