Welcome message sent when a client is registered
This introduces new components:

1. Communication BC (packaged as a gem)
2. Read model for displaying inbox

One new process manager:

- welcome message process

The Communication BC input is only SendMessage command and the output is MessageSent event.
The BC and the process manager have their own unit tests.
The UI is tested via integration test. In the future we will introduce unit tests for the read model.