Merged OrderItemInvoicingProcess into InvoiceGeneration process
Removed InvoiceItemCalculated event.
Removed the choreography with VAT determination via commands and events. Instead now, we access to the VatRateCatalog - not perfect either. Next time we work in this area, it would be nice to just subscribe in process for VAT events. Or have a read model for the process.
Many tests had to be updated to set proper vat rate.