This Little Maneuver's Gonna Cost Us 51 Years
Removed Pricing.event_store at the class level.
It was there when Sidekiq existed. Later, when we removed Sidekiq this hack wasn't cleared.
This code together with parralelism and mutant runner experiment resulted in 2(!) in-memory event_store to exist in the same test scenario.
It was super-hard to debug. Thanks to Claude for spotting it.