Make time promotions in tests active at different times
This is a hack to avoid solving the problem of global time promotions yet.
The problem is that we dont scope time promotions and as such across different parralel test runs they confuse next test scenarios.
Another hack would be to clear the in-memory repo between runs.

The proper solution seems to be to introduce the concept of stores. Then time promotions would be scoped per such store.