New Stores BC
First step towards multitenancy based on stores. Maybe in the future we will allow many stores per one "client" but for now we dont bother about it.
This BC ony registers a store for now. This opens the options to assign products/clients/orders/promotions to the store, once it's registered.