Store drop-down in UI
It ensures some store is always selected.
Because we sometimes reset production db - we select the first one if cookie doesnt match.
In the future it will be changed to select store by url/subdomain etc.

We should probably have stores read model for that later on - for now it's using Admin::Stores.